requirements.txt
requests
pyyaml
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
